[MWB] - moving Common.DragDrop.cs to DragDrop.cs - #35155
